+++ app_flutter/test/status_grid_test.dart
@@ -20,6 +20,8 @@ import 'package:app_flutter/status_grid.dart';
 import 'package:app_flutter/task_box.dart';
 import 'package:app_flutter/task_matrix.dart' show TaskMatrix;
 
+import 'utils/fake_flutter_build.dart';
+
 void main() {
   group('StatusGrid', () {
     List<CommitStatus> statuses;
@@ -229,143 +231,5 @@ void main() {
         }
       }
     });
-
-    testWidgets('error snackbar shown for error fetching tree status',
-        (WidgetTester tester) async {
-      final TestFlutterBuildState buildState = TestFlutterBuildState(
-        isTreeBuildingValue: CocoonResponse<bool>()..error = 'tree error',
-        statusesValue: CocoonResponse<List<CommitStatus>>()
-          ..data = <CommitStatus>[CommitStatus()],
-        hasError: true,
-      );
-      buildState.hasError = true;
-
-      await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-          body: Column(
-            children: <Widget>[
-              ChangeNotifierProvider<FlutterBuildState>.value(
-                value: buildState,
-                child: const StatusGridContainer(),
-              ),
-            ],
-          ),
-        ),
-      ));
-      await tester.pump(
-          const Duration(milliseconds: 750)); // open animation for snackbar
-
-      expect(
-          find.text(StatusGridContainer.errorFetchCommitStatus), findsNothing);
-      expect(
-          find.text(StatusGridContainer.errorFetchTreeStatus), findsOneWidget);
-
-      // Snackbar message should go away after its duration
-      await tester.pumpAndSettle(
-          StatusGridContainer.errorSnackbarDuration); // wait the duration
-      await tester.pump(); // schedule animation
-      await tester.pump(const Duration(milliseconds: 1500)); // close animation
-
-      buildState.hasError = false;
-
-      /// Update state to no longer have errors
-      buildState.notifyListeners();
-
-      // Wait another snackbar duration to prevent a race condition and ensure it clears
-      await tester.pumpAndSettle(StatusGridContainer.errorSnackbarDuration);
-
-      expect(
-          find.text(StatusGridContainer.errorFetchCommitStatus), findsNothing);
-      expect(find.text(StatusGridContainer.errorFetchTreeStatus), findsNothing);
-    });
-
-    testWidgets('error snackbar shown for error fetching commit statuses',
-        (WidgetTester tester) async {
-      final TestFlutterBuildState buildState = TestFlutterBuildState(
-        isTreeBuildingValue: CocoonResponse<bool>()..data = true,
-        statusesValue: CocoonResponse<List<CommitStatus>>()
-          ..data = <CommitStatus>[CommitStatus()]
-          ..error = 'statuses error',
-        hasError: true,
-      );
-
-      await tester.pumpWidget(MaterialApp(
-        home: Scaffold(
-          body: Column(
-            children: <Widget>[
-              ChangeNotifierProvider<FlutterBuildState>.value(
-                value: buildState,
-                child: const StatusGridContainer(),
-              ),
-            ],
-          ),
-        ),
-      ));
-
-      await tester.pump(
-          const Duration(milliseconds: 750)); // open animation for snackbar
-
-      expect(find.text(StatusGridContainer.errorFetchTreeStatus), findsNothing);
-      expect(find.text(StatusGridContainer.errorFetchCommitStatus),
-          findsOneWidget);
-
-      // Snackbar message should go away after its duration
-      await tester.pumpAndSettle(
-          StatusGridContainer.errorSnackbarDuration); // wait the duration
-      await tester.pump(); // schedule animation
-      await tester.pump(const Duration(milliseconds: 1500)); // close animation
-
-      buildState.hasError = false;
-
-      /// Update state to no longer have errors
-      buildState.notifyListeners();
-
-      // Wait another snackbar duration to prevent a race condition and ensure it clears
-      await tester.pumpAndSettle(StatusGridContainer.errorSnackbarDuration);
-
-      expect(
-          find.text(StatusGridContainer.errorFetchCommitStatus), findsNothing);
-      expect(find.text(StatusGridContainer.errorFetchTreeStatus), findsNothing);
-    });
   });
-}
-
-/// A [ChangeNotifier] for testing the interactions with error handling.
-class TestFlutterBuildState extends ChangeNotifier
-    implements FlutterBuildState {
-  TestFlutterBuildState(
-      {@required this.isTreeBuildingValue,
-      @required this.statusesValue,
-      this.hasError});
-
-  @override
-  bool hasError = false;
-
-  @override
-  CocoonResponse<bool> get isTreeBuilding => isTreeBuildingValue;
-  CocoonResponse<bool> isTreeBuildingValue;
-
-  @override
-  CocoonResponse<List<CommitStatus>> get statuses => statusesValue;
-  CocoonResponse<List<CommitStatus>> statusesValue;
-
-  @override
-  Timer refreshTimer;
-
-  @override
-  GoogleSignInService authService;
-
-  @override
-  Duration get refreshRate => null;
-
-  @override
-  Future<bool> rerunTask(Task task) => null;
-
-  @override
-  Future<void> signIn() => null;
-  @override
-  Future<void> signOut() => null;
-
-  @override
-  Future<void> startFetchingBuildStateUpdates() => null;
 }
